refactor: Test 수정
1. AdvanceJudgementTest 전진 불가능 경우 테스트에 @DisplayName 추가
2. CarNamesValidatorTest와 TryCountValidatorTest 통합
